GRF-777 : Moving InternalApiCategory.php DTO to Domain\Model\Enrichment
