+++ src/Akeneo/Category/back/tests/Integration/Domain/Model/Enrichment/InternalApiCategoryIntegration.php
@@ -2,12 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Category\back\tests\Integration\ServiceApi;
+namespace Akeneo\Category\back\tests\Integration\Domain\Model\Enrichment;
 
 use Akeneo\Category\back\tests\Integration\Helper\CategoryTestCase;
-use Akeneo\Category\Domain\Model\Enrichment\Category;
-use Akeneo\Category\ServiceApi\InternalApiCategory;
-use Webmozart\Assert\InvalidArgumentException;
+use Akeneo\Category\Domain\Model\Enrichment\InternalApiCategory;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
